% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth.R
\name{speciesGS}
\alias{speciesGS}
\title{Map Species Names}
\usage{
subsetGS(sp)
}
\arguments{
\item{sp}{set of species of interest (either binomial or short code name)}
}
\value{
A dataframe with the information for the requested species.
}
\description{
Map binomial species name to short code species name and vice versa
}
\details{
bla
}
\examples{
speciesGS(c("Pinus pinaster", "Ath"))
}
